requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
python-dotenv>=0.19.0
tqdm>=4.62.0
scikit-learn>=0.24.0